requirements.txt
fastapi==0.116.1
paramiko==4.0.0
uvicorn==0.35.0
PyYAML==6.0.2
Jinja2==3.1.6
python-multipart==0.0.20

